+++ xonsh/ptk/shell.py
@@ -58,11 +58,16 @@ class PromptToolkitShell(BaseShell):
         multicolumn = (completions_display == 'multi')
         self.styler.style_name = env.get('XONSH_COLOR_STYLE')
         completer = None if completions_display == 'none' else self.pt_completer
+        prompt_tokens = self.prompt_tokens(None)
+        get_prompt_tokens = lambda cli: prompt_tokens
+        rprompt_tokens = self.rprompt_tokens(None)
+        get_rprompt_tokens = lambda cli: rprompt_tokens
         with self.prompter:
             line = self.prompter.prompt(
                     mouse_support=mouse_support,
                     auto_suggest=auto_suggest,
-                    get_prompt_tokens=self.prompt_tokens,
+                    get_prompt_tokens=get_prompt_tokens,
+                    get_rprompt_tokens=get_rprompt_tokens,
                     style=PygmentsStyle(xonsh_style_proxy(self.styler)),
                     completer=completer,
                     lexer=PygmentsLexer(XonshLexer),
@@ -125,6 +130,20 @@ class PromptToolkitShell(BaseShell):
             print_exception()
         toks = partial_color_tokenize(p)
         self.settitle()
+        return toks
+
+    def rprompt_tokens(self, cli):
+        """Returns a list of (token, str) tuples for the current right
+        prompt.
+        """
+        p = builtins.__xonsh_env__.get('RIGHT_PROMPT')
+        if len(p) == 0:
+            return []
+        try:
+            p = partial_format_prompt(p)
+        except Exception:  # pylint: disable=broad-except
+            print_exception()
+        toks = partial_color_tokenize(p)
         return toks
 
     def format_color(self, string, **kwargs):